requirements.txt
gymnasium
numpy
neat-python
swig
imageio
graphviz
matplotlib